validate some raffle errors and fix types

+++ programs/raffle/src/state/tickets.rs
@@ -3,6 +3,6 @@ use anchor_lang::prelude::*;
 #[account]
 pub struct Tickets {
   pub raffle: Pubkey,
-  pub amount: u64,
+  pub amount: u32,
   pub bump: u8,
 }
